SBS-1032: Split the bound policy so rustc --test can prove it on Linux
Keep enqueue/dequeue free of the Tauri crate graph, matching managed_image,
and leave the async sender as a thin wrapper around that policy.
